# backend/event-service/event_openapi.yaml
openapi: 3.0.3
info:
  title: Event Service API
  description: API quản lý sự kiện và người dùng tham gia sự kiện trong hệ thống VolunteerHub
  version: 1.0.0
  contact:
    name: DongVu1105
    email: admin@gmail.com
servers:
  - url: http://localhost:8084/event
    description: Development server
...
paths:
  /create:
    post:
      tags:
        - Event Management
      summary: Tạo sự kiện mới
      description: Endpoint cho phép quản lý sự kiện tạo một sự kiện mới với thông tin chi tiết và hình ảnh (tùy chọn)
      operationId: createEvent
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - request
              properties:
                request:
                  type: string
                  description: JSON string chứa thông tin sự kiện
                  example: '{"title":"Sự kiện từ thiện","location":"Hà Nội","description":"Mô tả sự kiện","startDate":"2025-12-01T08:00:00Z","finishDate":"2025-12-02T17:00:00Z","category":"Từ thiện"}'
                file:
                  type: string
                  format: binary
                  description: File hình ảnh sự kiện (tùy chọn)
      responses:
        '200':
          description: Tạo sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEventResponse'
  /update:
    put:
      tags:
        - Event Management
      summary: Cập nhật sự kiện
      description: Endpoint cho phép quản lý sự kiện sửa đổi thông tin của một sự kiện đã tồn tại
      operationId: updateEvent
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - request
              properties:
                request:
                  type: string
                  description: JSON string chứa thông tin cập nhật sự kiện
                  example: '{"id":"event123","title":"Sự kiện từ thiện cập nhật","location":"Hà Nội","description":"Mô tả mới","startDate":"2025-12-01T08:00:00Z","finishDate":"2025-12-02T17:00:00Z","category":"Từ thiện"}'
                file:
                  type: string
                  format: binary
                  description: File hình ảnh sự kiện mới (tùy chọn)
      responses:
        '200':
          description: Cập nhật sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEventResponse'
  /category-date:
    get:
      tags:
        - Event Management
      summary: Lọc sự kiện theo danh mục và ngày
      description: Endpoint cho phép người dùng tìm kiếm và lọc sự kiện theo danh mục và khoảng thời gian
      operationId: findEventsByCategoryAndDate
      parameters:
        - name: page
          in: query
          description: Số trang (bắt đầu từ 1)
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: Số lượng sự kiện trên mỗi trang
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          description: Danh mục sự kiện
          schema:
            type: string
            default: Từ thiện
        - name: fromDate
          in: query
          description: Ngày bắt đầu lọc
          schema:
            type: string
            format: date-time
            default: '2000-01-01T08:00:00Z'
        - name: toDate
          in: query
          description: Ngày kết thúc lọc
          schema:
            type: string
            format: date-time
            default: '2100-01-01T08:00:00Z'
      responses:
        '200':
          description: Lấy danh sách sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageEventResponse'
...
  /status/{eventId}:
    get:
      tags:
        - Event Management
      summary: Kiểm tra trạng thái duyệt sự kiện
      description: Endpoint kiểm tra xem sự kiện đã được admin duyệt chưa để có thể đăng bài
      operationId: checkEventStatus
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID của sự kiện
          schema:
            type: string
      responses:
        '200':
          description: Lấy trạng thái sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
  /pending:
    get:
      tags:
        - Event Management
      summary: Tìm các sự kiện chưa được chấp nhận
      description: Endpoint cho admin tìm các sự kiện đang chờ duyệt
      operationId: findPendingEvents
      parameters:
        - name: page
          in: query
          description: Số trang (bắt đầu từ 1)
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: Số lượng sự kiện trên mỗi trang
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lấy danh sách sự kiện chờ duyệt thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageEventResponse'
  /accept/{eventId}:
    put:
      tags:
        - Event Management
      summary: Chấp nhận sự kiện
      description: Endpoint cho phép admin chấp nhận và duyệt một sự kiện
      operationId: acceptEvent
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID của sự kiện cần chấp nhận
          schema:
            type: string
      responses:
        '200':
          description: Chấp nhận sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEventResponse'
  /delete/{eventId}:
    delete:
      tags:
        - Event Management
      summary: Xóa sự kiện
      description: Endpoint cho phép admin xóa một sự kiện
      operationId: deleteEvent
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID của sự kiện cần xóa
          schema:
            type: string
      responses:
        '200':
          description: Xóa sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
components:
  schemas:
    ApiResponseEventResponse:
      type: object
      properties:
        code:
          type: integer
          default: 0
        message:
          type: string
          default: success!
        data:
          $ref: '#/components/schemas/EventResponse'
    ApiResponsePageEventResponse:
      type: object
      properties:
        code:
          type: integer
          default: 0
        message:
          type: string
          default: success!
        data:
          allOf:
            - $ref: '#/components/schemas/PageResponse'
            - type: object
              properties:
                result:
                  type: array
                  items:
                    $ref: '#/components/schemas/EventResponse'
    ApiResponseBoolean:
      type: object
      properties:
        code:
          type: integer
          default: 0
        message:
          type: string
          default: success!
        data:
          type: boolean
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          default: 0
        message:
          type: string
          default: success!
        data:
          type: object
          nullable: true
...
    EventResponse:
      type: object
      properties:
        id:
          type: string
          description: ID của sự kiện
        title:
          type: string
          description: Tiêu đề sự kiện
        location:
          type: string
          description: Địa điểm tổ chức sự kiện
        description:
          type: string
          description: Mô tả chi tiết về sự kiện
        startDate:
          type: string
          format: date-time
          description: Ngày giờ bắt đầu sự kiện
        finishDate:
          type: string
          format: date-time
          description: Ngày giờ kết thúc sự kiện
        category:
          type: string
          description: Danh mục sự kiện
        managerId:
          type: string
          description: ID của người quản lý sự kiện
        managerUsername:
          type: string
          description: Tên đăng nhập của người quản lý sự kiện
        createdDate:
          type: string
          description: Ngày tạo sự kiện
        image:
          type: string
          description: URL hình ảnh sự kiện
        statusEvent:
          type: boolean
          description: Trạng thái duyệt của sự kiện
...
    PageResponse:
      type: object
      properties:
        currentPage:
          type: integer
          description: Trang hiện tại
        totalPages:
          type: integer
          description: Tổng số trang
        pageSize:
          type: integer
          description: Số lượng phần tử trên mỗi trang
        totalElements:
          type: integer
          format: int64
          description: Tổng số phần tử
        hasNextPage:
          type: boolean
          description: Có trang tiếp theo hay không
        hasPreviousPage:
          type: boolean
          description: Có trang trước đó hay không
        result:
          type: array
          items:
            type: object
          description: Danh sách kết quả
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []