# backend/event-service/src/main/java/com/dongVu1105/event_service/controller/EventController.java
package com.dongVu1105.event_service.controller;

import com.dongVu1105.event_service.dto.ApiResponse;
import com.dongVu1105.event_service.dto.request.EventCreationRequest;
import com.dongVu1105.event_service.dto.request.EventUpdationRequest;
import com.dongVu1105.event_service.dto.response.EventResponse;
import com.dongVu1105.event_service.dto.response.PageResponse;
import com.dongVu1105.event_service.service.EventService;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import jakarta.validation.Valid;
import lombok.AccessLevel;
import lombok.RequiredArgsConstructor;
import lombok.experimental.FieldDefaults;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.time.Instant;
import java.util.List;

@RestController
@RequiredArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
public class EventController {
    EventService eventService;
    ObjectMapper objectMapper;
    RequestValidator requestValidator;

    // Quản lí sự kiện tạo event
    @PostMapping("/create")
    public ApiResponse<EventResponse> create (
            @Valid
            @RequestPart(value = "request") String requestJson,
            @RequestPart(value = "file", required = false)MultipartFile file) throws JsonProcessingException {
        EventCreationRequest request = objectMapper.readValue(requestJson, EventCreationRequest.class);
        requestValidator.validate(request);
        return ApiResponse.<EventResponse>builder().data(eventService.create(request, file)).build();
    }

    // Quản lí sự kiện sửa event
    @PutMapping("/update")
    public ApiResponse<EventResponse> update (
            @RequestPart(value = "request") String requestJson,
            @RequestPart(value = "file", required = false) MultipartFile file) throws JsonProcessingException {
        EventUpdationRequest request = objectMapper.readValue(requestJson, EventUpdationRequest.class);
        requestValidator.validate(request);
        return ApiResponse.<EventResponse>builder().data(eventService.update(request, file)).build();
    }

    // Người dùng lọc event theo danh mục và ngày
    @GetMapping("/category-date")
    public ApiResponse<PageResponse<EventResponse>> findAllByCategoryAndDate (
            @RequestParam(value = "page", required = false, defaultValue = "1") int page,
            @RequestParam(value = "size", required = false, defaultValue = "10") int size,
            @RequestParam(value = "category", required = false, defaultValue = "Từ thiện") String category,
            @RequestParam(value = "fromDate", required = false, defaultValue = "2000-01-01T08:00:00Z") Instant fromDate,
            @RequestParam(value = "toDate", required = false, defaultValue = "2100-01-01T08:00:00Z") Instant toDate){
        return ApiResponse.<PageResponse<EventResponse>>builder()
                .data(eventService.findAllByCategoryAndDate(page, size, category, fromDate, toDate)).build();
    }

    // Tìm event theo id
    @GetMapping("/{eventId}")
    public ApiResponse<EventResponse> findById (@PathVariable("eventId") String eventId){
        return ApiResponse.<EventResponse>builder().data(eventService.findById(eventId)).build();
    }

    // Kiểm tra trạng thái event (admin duyệt chưa)
    @GetMapping("/status/{eventId}")
    public ApiResponse<Boolean> ableToPost (@PathVariable("eventId") String eventId){
        return ApiResponse.<Boolean>builder().data(eventService.ableToPost(eventId)).build();
    }

    // Admin tìm các sự kiện chưa được chấp nhận
    @GetMapping("/pending")
    public ApiResponse<PageResponse<EventResponse>> findAllPendingEvent (
            @RequestParam(value = "page", required = false, defaultValue = "1") int page,
            @RequestParam(value = "size", required = false, defaultValue = "10") int size){
        return ApiResponse.<PageResponse<EventResponse>>builder()
                .data(eventService.findAllPendingEvent(page, size)).build();
    }

    // Admin chấp nhận event
    @PutMapping("/accept/{eventId}")
    public ApiResponse<EventResponse> acceptEvent (@PathVariable("eventId") String eventId){
        return ApiResponse.<EventResponse>builder().data(eventService.acceptEvent(eventId)).build();
    }

    // Admin xóa event
    @DeleteMapping("/delete/{eventId}")
    public ApiResponse<Void> delete (@PathVariable("eventId") String eventId){
        eventService.delete(eventId);
        return ApiResponse.<Void>builder().build();
    }




}